backend/llm-service/pyproject.toml
[tool.poetry]
name = "llm-service"
version = "0.1.0"
description = ""
authors = ["JangSeHun <westreed@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.124.0"
uvicorn = "^0.38.0"
python-dotenv = "^1.2.1"
openai = "^2.9.0"
py-eureka-client = "^0.11.10"
requests = "2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
